requirements.txt
python-telegram-bot
psycopg2-binary
telethon
pika
pytz
tzlocal
pySocks
gitpython
